Initial Commit with code
This is the core code for the app testing repository. It contains a
local version of the tf-mod-shared module, a terraform validate GitHub
Action workflow, a terraform plan GitHub Action workflow, and the
associated scripts for those workflows. The apps directory is currently
empty.

The backend.hcl file in the top directory is configured for
remote state stored in S3 in our Dev/Test environment.