working code for testing
This is the working code to build the S3 bucket for the ETD-83
Jira task. The dev.tfvars have been pushed to Parameter Store in the
Dev/Test environment, so running the aws_parse.py properly will
generate the correct .tfvars file for the dev environment.

The next step is to review the versioning and expiration settings with
Jeremy to ensure that it meets the criteria. Once this is fully tested
we can copy this code to mitlib-terraform and deploy in stage for
final testing.